+++ src/contentScript.ts
@@ -1,38 +1,56 @@
 console.log("Content script loaded.");
 
-const customStyle = document.createElement('style');
+const customStyle = document.createElement("style");
 document.head.appendChild(customStyle);
 
-let messageMaxWidthStyle = '';
-let messagePaddingStyle = ''; 
-let messageBorderRadiusStyle = '';
-let inputBoxMaxWidthStyle = '';
-let messageBoxColors = '';
-let selectionColors = '';
+let messageMaxWidthStyle = "";
+let messagePaddingStyle = "";
+let messageBorderRadiusStyle = "";
+let inputBoxMaxWidthStyle = "";
+let messageBoxColors = "";
+let selectionColors = "";
 let chatMessageButtons = `
     [data-testid] button {
       visibility: unset
     }
 `;
 
-const updateMessageColor = (color: string, isUser: boolean, isDark: boolean) => {
+const updateMessageColor = (
+    color: string,
+    isUser: boolean,
+    isDark: boolean,
+) => {
     messageBoxColors = `
       .dark {
-        [data-testid]:nth-child(even) > * > * { background-color: ${isUser && isDark ? color : '#4e7645'} }
-        [data-testid]:nth-child(odd) > * > * { background-color: ${!isUser && isDark ? color : '#3c6083'} }
+        [data-testid]:nth-child(even) > * > * { background-color: ${
+            isUser && isDark ? color : "#4e7645"
+        } }
+        [data-testid]:nth-child(odd) > * > * { background-color: ${
+            !isUser && isDark ? color : "#3c6083"
+        } }
       }
       .light {
-        [data-testid]:nth-child(even) > * > * { background-color: ${isUser && !isDark ? color : '#62B1F6'} }
-        [data-testid]:nth-child(odd) > * > * { background-color: ${!isUser && !isDark ? color : '#EEEEEE'} }
+        [data-testid]:nth-child(even) > * > * { background-color: ${
+            isUser && !isDark ? color : "#62B1F6"
+        } }
+        [data-testid]:nth-child(odd) > * > * { background-color: ${
+            !isUser && !isDark ? color : "#EEEEEE"
+        } }
       }
     `;
     updateAllStyles();
 };
 
 const updateAllStyles = () => {
-    customStyle.textContent = messageBoxColors + messageMaxWidthStyle + messagePaddingStyle 
-                              + messageBorderRadiusStyle + inputBoxMaxWidthStyle + selectionColors + chatMessageButtons;
-}
+    customStyle.textContent =
+        messageBoxColors +
+        messageMaxWidthStyle +
+        messagePaddingStyle +
+        messageBorderRadiusStyle +
+        inputBoxMaxWidthStyle +
+        selectionColors +
+        chatMessageButtons;
+};
 
 const updateMessageMaxWidth = (widthPercentage: number) => {
     messageMaxWidthStyle = `@media (min-width: 1200px) { [data-testid] > * > * { max-width: ${widthPercentage}% } }`;
@@ -54,10 +72,10 @@ const updateInputBoxMaxWidth = (widthPercentage: number) => {
       form { max-width: ${widthPercentage}% !important; }
     }`;
     updateAllStyles();
-}
+};
 
 const resetDefaultMessageColors = () => {
-  messageBoxColors = `
+    messageBoxColors = `
   .dark {
     [data-testid]:nth-child(even) > * > * { background-color: #4e7645 }
     [data-testid]:nth-child(odd) > * > * { background-color: #3c6083 }
@@ -67,30 +85,71 @@ const resetDefaultMessageColors = () => {
     [data-testid]:nth-child(odd) > * > * { background-color: #EEEEEE }
   }
 `;
-  updateAllStyles();
+    updateAllStyles();
+};
+
+const $main = document.querySelector("main");
+
+if ($main) {
+    const scrollToBottom = () => {
+        const $messagesContainer = document.querySelector(
+            '[role="presentation"] > div > div > div',
+        );
+        if ($messagesContainer) {
+            setTimeout(
+                () =>
+                    ($messagesContainer.scrollTop =
+                        $messagesContainer.scrollHeight),
+                0,
+            );
+            $messagesContainer.scrollTop = $messagesContainer.scrollHeight;
+        }
+    };
+
+    $main.addEventListener("keydown", (event: KeyboardEvent) => {
+        if (
+            event.target instanceof HTMLTextAreaElement &&
+            event.target.getAttribute("id") === "prompt-textarea" &&
+            event.target.textContent !== "" &&
+            event.key === "Enter" &&
+            !event.shiftKey
+        ) {
+            scrollToBottom();
+        }
+    });
+
+    $main.addEventListener("click", (event) => {
+        if (
+            (event.target instanceof HTMLButtonElement &&
+                event.target.getAttribute("data-testid") === "send-button") ||
+            (event.target instanceof SVGElement &&
+                event.target.closest('[data-testid="send-button"]'))
+        ) {
+            scrollToBottom();
+        }
+    });
 }
 
 const setDefaultSettings = () => {
     updateMessageMaxWidth(95);
-    updateMessagePadding('10px');
-    updateMessageBorderRadius('5px');
+    updateMessagePadding("10px");
+    updateMessageBorderRadius("5px");
     updateInputBoxMaxWidth(70);
     resetDefaultMessageColors();
-}
+};
 
 // set default settings on load
 setDefaultSettings();
 
 // listening for messages from the background script or popup
-chrome.runtime.onMessage.addListener(
-  (request, sender, sendResponse) => {
-    if (request.message === 'change_color') {
-      document.body.style.backgroundColor = request.color;
-      sendResponse({status: 'Color changed'});
+chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
+    if (request.message === "change_color") {
+        document.body.style.backgroundColor = request.color;
+        sendResponse({ status: "Color changed" });
     }
 });
 
 // send a message to the background script if needed
-chrome.runtime.sendMessage({ message: 'Content script active' }, response => {
-  console.log(response.reply);
+chrome.runtime.sendMessage({ message: "Content script active" }, (response) => {
+    console.log(response.reply);
 });